requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.1
pytz==2023.3
gunicorn==21.2.0